Issue 3130: Disk sizes must be 'long'

+++ src/main/java/icecube/daq/secBuilder/SecBuilderMonitor.java
@@ -87,7 +87,7 @@ public class SecBuilderMonitor implements SecBuilderMonitorMBean {
      *
      * @return the number of units still available in the disk.
      */
-    public int getDiskAvailable(){
+    public long getDiskAvailable(){
         return dispatcher.getDiskAvailable();
     }
 
@@ -97,7 +97,7 @@ public class SecBuilderMonitor implements SecBuilderMonitorMBean {
      *
      * @return the total number of units in the disk.
      */
-    public int getDiskSize(){
+    public long getDiskSize(){
         return dispatcher.getDiskSize();
     }
 }
