+++ src/main/java/icecube/daq/secBuilder/SecBuilderMonitorMBean.java
@@ -41,7 +41,7 @@ public interface SecBuilderMonitorMBean {
      *
      * @return the number of units still available in the disk.
      */
-    int getDiskAvailable();
+    long getDiskAvailable();
 
     /**
      * Returns the total number of units in the disk (measured in MB).
@@ -49,5 +49,5 @@ public interface SecBuilderMonitorMBean {
      *
      * @return the total number of units in the disk.
      */
-    int getDiskSize();
+    long getDiskSize();
 }
